Update minipack.js
//line 246

const graph = createGraph(
  path.join(__dirname, '../example/entry.js')
);

This line constructs an absolute path to the entry file of the application by joining the directory of the current file (__dirname) with the relative path to example/entry.js. Using __dirname ensures that the entry file is resolved correctly regardless of the directory from which the script is executed.